+++ components/debug/debugging.cpp
@@ -4,6 +4,34 @@
 
 namespace Debug
 {
+#ifdef _WIN32
+    bool attachParentConsole()
+    {
+        if (GetConsoleWindow() != nullptr)
+            return true;
+
+        if (AttachConsole(ATTACH_PARENT_PROCESS))
+        {
+            fflush(stdout);
+            fflush(stderr);
+            std::cout.flush();
+            std::cerr.flush();
+
+            // this looks dubious but is really the right way
+            _wfreopen(L"CON", L"w", stdout);
+            _wfreopen(L"CON", L"w", stderr);
+            _wfreopen(L"CON", L"r", stdin);
+            freopen("CON", "w", stdout);
+            freopen("CON", "w", stderr);
+            freopen("CON", "r", stdin);
+
+            return true;
+        }
+
+        return false;
+    }
+#endif
+
     std::streamsize DebugOutputBase::write(const char *str, std::streamsize size)
     {
         // Skip debug level marker
