--- components/debug/win32.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#ifdef _WIN32
-
-namespace Debug {
-    bool attachParentConsole();
-}
-
-#endif
